parsemail: Process email with duplicated message-id
Current email parsing filters-out messages with a message-id
that already exists in db, producing that some valid patch
revisions are not processed in patchwork.

This change allows for valid patches with duplicated message-id to be
processed by modifying the message-id in patchwork only when email content
in new message is different from that in the existing patch.

[YOCTO #10756]

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>