parsemail: Split PATCH prefix if not single
PATCH prefix is not appropriately identified as such if it is not
sepparated from following word in an email Subject line. This leads to
wrong email parsing and revisions not appended to initial series.

This change allows appropriate email parsing by checking if PATCH prefix
in Subject line is not sepparated from following word and adding a space
between them.

[YOCTO #10823]

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>