series.py: return bundle objects only when authenticated
If a non-authenticated user tries to get a series view from
patchwork web UI, a 500 error message is displayed caused by
the database access attempted when filtering the bundle objects.

This change allows not-logged users to acces series view by
returning an empty bundle variable, instead of trying to
access the database.

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>