parsemail.py: Improve status-change-through-email
Patch status can be edited through email by project maintainers only
without any feedback if this is attempted by any other user,
providing a poor user experience.

This change extends the patch-status-through-email functionality
to be performed by the series submitter also, and provides a
system-generated message if an attempt was made through a message
using an email address not recognized neither as project maintainer or
series submitter in patchwork.

[YOCTO #11027]

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>