parsemail.py: Improve clean_subject function
PATCH prefix is not appropriately identified as such if it is not
sepparated from following word in an email Subject line. This leads to
wrong patch naming.

This change allows for appropriate patch naming by checking during
clean_subject function if "PATCH" prefix in Subject line is not
sepparated from following word and adding a space between them.

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>