parsemail: keep branch name in subject for series
patches with different branch names included in their name are
wrongly appended as succesive revisions of the same patch in a
series. This is due to the patches being assigned to the same
series after their prefixes got removed during series naming/search.

This change looks for an updated release name list from the Yocto
wiki and keeps the release name in the series naming/search,
effectively creating individual series for patches targeted to
diferent releases.

[YOCTO #10716]

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>