htdocs/css/style.css: fix select style in FF and IE
Series view displays the "Revision" select element not in
line with the general patchwork style/design when viewed
using Firefox or Internet Explorer browsers due to a known
incompatibility of this browsers with Boostrap library
(for example, an arrow head inside a button is displayed next
to the "Revision" select element when using FF).

This change hides the out-of-style items by using browser-specific
css properties "moz-appearance" and "ms-expand".

[YOCTO #11886]

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>