patchwork/templates/patchwork/series.html: Add link column
Clicking on a patch name in Series view displays the patch at
the botton section in same page, but there is no easy way to
get a direct link to the patch, for example: to share the link.

This change includes the column "Link" where a direct link to the
patch is displayed and labeled with the patch id, that can be copied
or clicked to display the patch detail on a new tab.

[YOCTO #11888]

Signed-off-by: Jose Lamego <jose.a.lamego@linux.intel.com>